requirements.txt
PyYAML>=6.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pypdf>=4.2.0
docling>=2.84.0
sentence-transformers>=2.7.0
chromadb>=0.5.0
requests>=2.32.0
openai>=1.40.0
pandas>=2.2.0
SQLAlchemy>=2.0.0
pyodbc>=5.1.0
